refactor: extract participants async job worker into conversion_participants_convert.py
